Moved ts modules into a separate directory
