+++ tests/launch.spec.ts
@@ -14,8 +14,8 @@
  *  limitations under the License.
  */
 import {test, expect} from '@playwright/test';
-import {Terra} from './terra';
-import {TimeUnits} from './timeunits';
+import {Terra} from '../modules/terra';
+import {TimeUnits} from '../modules/timeunits';
 
 require('dotenv').config({path: __dirname + '/.env.local'})
