+++ tests/pause.spec.ts
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 import {test, expect} from '@playwright/test';
-import {Terra} from './terra';
+import {Terra} from '../modules/terra';
 
 test.describe('shutdown the instance but keep the disks', () => {
     test('Delete the Galaxy instance', async ({page}) => {
